# server/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "HTTP API for Phase 1 user top-up campaigns (admin + user-facing). Operates under `/campaign-center-api/v1`;",
        "title": "Campaign Center API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/campaign-center-api/v1",
    "paths": {
...
        "/web/user-profile": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user-profile"
                ],
                "summary": "Get user profile (user)",
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/api.UserProfileHTTPResponse"
                        }
                    },
                    "404": {
                        "description": "user not found",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    },
                    "503": {
                        "description": "database unavailable",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "api.UserProfileData": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "a***e@example.com"
                },
                "kycChecked": {
                    "type": "boolean",
                    "example": true
                },
                "registeredAt": {
                    "type": "string",
                    "example": "2026-05-16T10:00:00Z"
                },
                "username": {
                    "type": "string",
                    "example": "alice"
                }
            }
        },
        "api.UserProfileHTTPResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "$ref": "#/definitions/api.UserProfileData"
                },
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "data.StandardResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        }
    }
}